# src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: cmp是一个客户关系管理平台
  version: 1.0.0
  title: 客户关系管理平台
basePath: /cam
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义接口数据
paths:
...
  /sys/schedule/list:
    get:
      tags:
        - 定时任务
      summary: 定时任务列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: beanName
          description: spring bean名称
          in: query
          type: string
      responses:
        '200':
          description: 返回定时任务列表
          schema:
            $ref: '#/definitions/ScheduleJobEntityList'
  /sys/schedule/info/{jobId}:
    get:
      tags:
        - 定时任务
      summary: 获取定时任务信息
      produces:
        - application/json
      parameters:
        - name: jobId
          description: 定时任务ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回定时任务信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              schedule:
                description: 定时任务信息
                $ref: '#/definitions/ScheduleJobEntity'
  /sys/schedule/save:
    post:
      tags:
        - 定时任务
      summary: 添加定时任务
      produces:
        - application/json
      parameters:
        - name: body
          description: 定时任务对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/ScheduleJobEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/schedule/update:
    post:
      tags:
        - 定时任务
      summary: 修改定时任务
      produces:
        - application/json
      parameters:
        - name: body
          description: 定时任务对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/ScheduleJobEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/schedule/delete:
    post:
      tags:
        - 定时任务
      summary: 删除定时任务
      produces:
        - application/json
      parameters:
        - name: body
          description: 定时任务ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/schedule/run:
    post:
      tags:
        - 定时任务
      summary: 立即执行任务
      produces:
        - application/json
      parameters:
        - name: body
          description: 定时任务ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/schedule/pause:
    post:
      tags:
        - 定时任务
      summary: 暂停定时任务
      produces:
        - application/json
      parameters:
        - name: body
          description: 定时任务ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/schedule/resume:
    post:
      tags:
        - 定时任务
      summary: 恢复定时任务
      produces:
        - application/json
      parameters:
        - name: body
          description: 定时任务ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /cus/datatraceability/list:
    get:
      tags:
        - 数据追溯管理
      summary: 数据追溯列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: dataName
          description: 数据名称
          in: query
          type: string
        - name: dataAttribute
          description: 数据字段属性
          in: query
          type: string
        - name: displayName
          description: 显示名称
          in: query
          type: string
        - name: type
          description: 表类型
          in: query
          type: string
        - name: dataValue
          description: 数据值
          in: query
          type: string
        - name: updateCode
          description: 修改人代码
          in: query
          type: string
        - name: updateName
          description: 修改人
          in: query
          type: string
        - name: startUpdateTime
          description: 开始修改时间
          in: query
          type: string
          format: date
        - name: endUpdateTime
          description: 结束修改时间
          in: query
          type: string
          format: date
      responses:
        '200':
          description: 返回数据追溯列表
          schema:
            $ref: '#/definitions/DataTraceabilityEntityList'
  /cus/datatraceability/info/{id}:
    get:
      tags:
        - 数据追溯管理
      summary: 获取数据追溯信息
      produces:
        - application/json
      parameters:
        - name: id
          description: ID
          in: path
          type: string
          required: true
      responses:
        '200':
          description: 返回数据追溯信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              user:
                $ref: '#/definitions/DataTraceabilityEntity'
  /cus/datatraceability/save:
    post:
      tags:
        - 数据追溯管理
      summary: 添加数据追溯
      produces:
        - application/json
      parameters:
        - name: body
          description: 数据追溯对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/DataTraceabilityEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /cus/datatraceability/update:
    post:
      tags:
        - 数据追溯管理
      summary: 修改数据追溯
      produces:
        - application/json
      parameters:
        - name: body
          description: 数据追溯对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/DataTraceabilityEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /cus/datatraceability/delete:
    post:
      tags:
        - 数据追溯管理
      summary: 删除数据追溯
      produces:
        - application/json
      parameters:
        - name: body
          description: 数据追溯ID列表
          in: body
          type: array
          items:
            type: string
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /cus/customertrajectory/list:
    get:
      tags:
        - 客户轨迹管理
      summary: 客户轨迹列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: customerId
          description: 客户id
          in: query
          type: string
        - name: trajectoryName
          description: 轨迹名称
          in: query
          type: string
        - name: trajectoryType
          description: 轨迹类型
          in: query
          type: string
        - name: createCode
          description: 创建人代码
          in: query
          type: string
        - name: createName
          description: 创建人
          in: query
          type: string
        - name: startCreateTime
          description: 开始创建时间
          in: query
          type: string
          format: date
        - name: endCreateTime
          description: 结束创建时间
          in: query
          type: string
          format: date
        - name: startCreateTime
          description: 开始创建时间
          in: query
          type: string
          format: date
        - name: endCreateTime
          description: 结束创建时间
          in: query
          type: string
          format: date
      responses:
        '200':
          description: 返回客户轨迹列表
          schema:
            $ref: '#/definitions/CustomerTrajectoryEntityList'
  /cus/customertrajectory/info/{id}:
    get:
      tags:
        - 客户轨迹管理
      summary: 获取客户轨迹信息
      produces:
        - application/json
      parameters:
        - name: id
          description: ID
          in: path
          type: string
          required: true
      responses:
        '200':
          description: 返回客户轨迹信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              user:
                $ref: '#/definitions/CustomerTrajectoryEntity'
  /cus/customertrajectory/save:
    post:
      tags:
        - 客户轨迹管理
      summary: 添加客户轨迹
      produces:
        - application/json
      parameters:
        - name: body
          description: 客户轨迹对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/CustomerTrajectoryEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /cus/customertrajectory/update:
    post:
      tags:
        - 客户轨迹管理
      summary: 修改客户轨迹
      produces:
        - application/json
      parameters:
        - name: body
          description: 客户轨迹对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/CustomerTrajectoryEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /cus/customertrajectory/delete:
    post:
      tags:
        - 客户轨迹管理
      summary: 删除客户轨迹
      produces:
        - application/json
      parameters:
        - name: body
          description: 客户轨迹ID列表
          in: body
          type: array
          items:
            type: string
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /cus/customertrajectoryson/list:
    get:
      tags:
        - 客户轨迹子表管理
      summary: 客户轨迹子表列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: trajectoryId
          description: 轨迹id
          in: query
          type: string
        - name: businessId
          description: 业务id
          in: query
          type: string
        - name: businessName
          description: 业务名称
          in: query
          type: string
      responses:
        '200':
          description: 返回客户轨迹子表列表
          schema:
            $ref: '#/definitions/CustomerTrajectorySonEntityList'
  /cus/customertrajectoryson/info/{id}:
    get:
      tags:
        - 客户轨迹子表管理
      summary: 获取客户轨迹子表信息
      produces:
        - application/json
      parameters:
        - name: id
          description: ID
          in: path
          type: string
          required: true
      responses:
        '200':
          description: 返回客户轨迹子表信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              user:
                $ref: '#/definitions/CustomerTrajectorySonEntity'
  /cus/customertrajectoryson/save:
    post:
      tags:
        - 客户轨迹子表管理
      summary: 添加客户轨迹子表
      produces:
        - application/json
      parameters:
        - name: body
          description: 客户轨迹子表对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/CustomerTrajectorySonEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /cus/customertrajectoryson/update:
    post:
      tags:
        - 客户轨迹子表管理
      summary: 修改客户轨迹子表
      produces:
        - application/json
      parameters:
        - name: body
          description: 客户轨迹子表对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/CustomerTrajectorySonEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /cus/customertrajectoryson/delete:
    post:
      tags:
        - 客户轨迹子表管理
      summary: 删除客户轨迹子表
      produces:
        - application/json
      parameters:
        - name: body
          description: 客户轨迹子表ID列表
          in: body
          type: array
          items:
            type: string
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
      type: object
      properties:
        code:
          description: 状态码 0：成功  非0：失败
          type: integer
          format: int32
        msg:
          description: 失败原因
          type: string
...
  ScheduleJobEntity:
    type: object
    properties:
      jobId:
        description: 任务ID
        type: integer
        format: int64
      beanName:
        description: spring bean名称
        type: string
      methodName:
        description: 方法名
        type: string
      params:
        description: 参数
        type: string
      cronExpression:
        description: cron表达式
        type: string
      status:
        description: 任务状态  0：正常  1：暂停
        type: integer
        format: int32
      remark:
        description: 备注
        type: string
      createTime:
        description: 创建时间
        type: string
        format: date-time
  ScheduleJobEntityList:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/ScheduleJobEntity'
...
  DataTraceabilityEntity:
    type: object
    properties:
      id:
        description: ID
        type: string
      dataName:
        description: 数据名称
        type: string
      dataAttribute:
        description: 数据字段属性
        type: string
      displayName:
        description: 显示名称
        type: string
      type:
        description: 表类型
        type: string
      dataValue:
        description: 数据值
        type: string
      updateCode:
        description: 修改人代码
        type: string
      updateName:
        description: 修改人
        type: string
      updateTime:
        description: 修改时间
        type: string
  DataTraceabilityEntityList:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/DataTraceabilityEntity'
  CustomerTrajectoryEntity:
    type: object
    properties:
      id:
        description: ID
        type: string
      customerId:
        description: 客户id
        type: string
      trajectoryName:
        description: 轨迹名称
        type: string
      trajectoryContent:
        description: 轨迹内容
        type: string
      trajectoryType:
        description: 轨迹类型
        type: string
      createCode:
        description: 创建人代码
        type: string
      createName:
        description: 创建人
        type: string
      createTime:
        description: 创建时间
        type: string
      updateCode:
        description: 修改人代码
        type: string
      updateName:
        description: 修改人
        type: string
      updateTime:
        description: 修改时间
        type: string
      customerTrajectorySonList:
        description: 客户轨迹子表列表
        type: array
        items:
          $ref: '#/definitions/CustomerTrajectorySonEntity'
  CustomerTrajectoryEntityList:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/CustomerTrajectoryEntity'
  CustomerTrajectorySonEntity:
    type: object
    properties:
      id:
        description: ID
        type: string
      trajectoryId:
        description: 轨迹id
        type: string
      businessId:
        description: 业务id
        type: string
      businessName:
        description: 业务名称
        type: string
      businessUrl:
        description: 业务url
        type: string
      createCode:
        description: 创建人代码
        type: string
      createName:
        description: 创建人
        type: string
      createTime:
        description: 创建时间
        type: string
      updateCode:
        description: 修改人代码
        type: string
      updateName:
        description: 修改人
        type: string
      updateTime:
        description: 修改时间
        type: string
  CustomerTrajectorySonEntityList:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/CustomerTrajectorySonEntity'
...